// src/models/CompanyUserProfileLanguageModel.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Link } from './Link';
import type { ProfileLanguageModel } from './ProfileLanguageModel';
export type CompanyUserProfileLanguageModel = {
    profileId?: number | null;
    language?: ProfileLanguageModel | null;
    level?: number | null;
    companyId?: number | null;
    companyUserId?: number | null;
    id?: number | null;
    url?: string | null;
    links?: Array<Link> | null;
};